requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.18
Werkzeug==0.9.4
argparse==1.2.1
itsdangerous==0.23
pyusb==1.0.0b1
wsgiref==0.1.2
